spectart-ai-service/requirements.txt
fastapi
uvicorn[standard]
openai>=1.37.0
pydantic
python-dotenv